requirements.txt
openai
streamlit
pandas
fuzzywuzzy
python-dotenv